refactor: FriendService CQS 패턴 적용
